+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to
 You can also check the
 [release page](https://github.com/visualize-admin/electricity-prices-switzerland/releases)
 
+# Unreleased
+
+- Scripts to perform GraphQL endpoint comparisons
+
 # 2.23.0 - 2025-11-06
 
 - Ability to switch SPARQL endpoint for the current user through a password protected interface